disable invoiceModule

+++ src/main/java/com/smartshop/eshop/repository/ShippingOriginRepository.java
@@ -8,5 +8,5 @@ import com.smartshop.eshop.domain.ShippingOrigin;
 public interface ShippingOriginRepository extends JpaRepository<ShippingOrigin, Long> {
 
 	@Query("select s from ShippingOrigin as s join fetch s.merchantStore sm where sm.id = ?1")
-	ShippingOrigin findByStore(Integer storeId);
+	ShippingOrigin findByStore(Long storeId);
 }
